changes made based on review, waiting for a few replies
