+++ src/web/lib/hakuneko/frontend@classic-light/chapters.html
@@ -439,10 +439,6 @@
                 Engine.ChaptermarkManager.addChaptermark( this.selectedChapter );
             }
 
-            onRequestHideViewer ( e ) {
-                this.set('selectedChapter', e.detail);
-                Engine.ChaptermarkManager.addChaptermark( this.selectedChapter );
-            }
             /**
              *
              */
