+++ src/web/mjs/engine/Settings.mjs
@@ -40,7 +40,7 @@ export default class Settings {
             docs = '.';
         }
         //HYPOFLEX
-        this.fs = require( 'fs' );
+        this.fs = require( 'fs-extra' );
 
         this.frontend = {
             label: 'Frontend ⁽¹⁾',
@@ -335,41 +335,5 @@ export default class Settings {
             return value;
 
         }
-    }
-
-    /**
-     * HYPOFLEX
-     * Save the given value for the given key in the bookmark storage
-     */
-    saveBookmarks( key, value, indentation ) {
-        return new Promise( ( resolve, reject ) => {
-            this.fs.writeFile( this.bookmarkDirectory.value + '\\hakuneko.' + key, JSON.stringify( value, undefined, indentation ), function( error ) {
-                if( error ) {
-                    reject( error );
-                } else {
-                    resolve();
-                }
-            } );
-        } );
-    }
-
-    /**
-     * HYPOFLEX
-     * Load the value for the given key from the bookmark storage
-     */
-    async loadBookmarks( key ) {
-        //return fetch( this.config + key ).then( response => response.json() );
-        return new Promise( ( resolve, reject ) => {
-            this.fs.readFile( this.bookmarkDirectory.value + '\\hakuneko.' + key, 'utf8', ( error, data ) => {
-                try {
-                    if( error ) {
-                        throw error;
-                    }
-                    resolve( JSON.parse( data ) );
-                } catch( e ) {
-                    reject( e );
-                }
-            } );
-        } );
     }
 }
